backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.2
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-multipart==0.0.9
pydantic-settings==2.3.0
alembic==1.13.1